v1.0a
The TwitterAPI Class has been imported to the GUI (Main class).
New function was added in the TwitterAPI class where it checks if the Tweet surpasses the character limit (280 in this case).
The tweet function has been imported to the Main class.

TODO: Make the GUI return the errors, put a lockdown to the Tweet button for 16 minutes*, clean up some code, and final release!

*I'm studying the Twitter API and as far as I know, 16 minutes is a good choice, but I am not sure yet so I will keep studying it.